refactor: 연관된 도메인 별 패키지 이동
