+++ src/main/java/domain/LadderGameResult.java
@@ -1,7 +1,13 @@
 package domain;
 
+import domain.exception.ExceptionType;
+import domain.exception.LadderGameException;
 import domain.ladder.Ladder;
 import domain.ladder.LadderPositions;
+import domain.name.Name;
+import domain.name.Names;
+import domain.result.LadderResult;
+import domain.result.LadderResults;
 import java.util.Map;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
