+++ src/main/java/domain/exception/LadderGameException.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.exception;
 
 public class LadderGameException extends RuntimeException {
     public LadderGameException(ExceptionType exceptionType) {
