+++ src/test/java/domain/LadderGameTest.java
@@ -2,6 +2,7 @@ package domain;
 
 import domain.ladder.Bridge;
 import domain.ladder.BridgeGenerator;
+import domain.name.Name;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
