+++ src/test/java/domain/ladder/RowTest.java
@@ -1,7 +1,7 @@
 package domain.ladder;
 
-import domain.ExceptionType;
-import domain.LadderGameException;
+import domain.exception.ExceptionType;
+import domain.exception.LadderGameException;
 import java.util.List;
 import java.util.stream.IntStream;
 import org.assertj.core.api.Assertions;
