+++ src/test/java/domain/result/LadderResultsTest.java
@@ -1,5 +1,7 @@
-package domain;
+package domain.result;
 
+import domain.exception.ExceptionType;
+import domain.exception.LadderGameException;
 import java.util.Arrays;
 import java.util.List;
 import org.assertj.core.api.Assertions;
